Changed render to redirect after a Create Update Delete
Previously it would render a new page, this caused issues and required a function. This function has been removed as the redirect will run the read command as if navigating. This also helps as refreshing the browser would go to the api address which would not work. This works better for the user.